Mayonnaise is also an instrument

+++ GXPEngine/Stands/CheeseStand.cs
@@ -7,7 +7,7 @@ using GXPEngine;
 
 public class CheeseStand : MarketStand
 {
-    public CheeseStand(float givenX, float givenY, int givenRotation = 0) : base(givenX, givenY, "cheeseStand.png", givenRotation)
+    public CheeseStand(float givenX, float givenY, float givenRotation = 0) : base(givenX, givenY, "cheeseStand.png", givenRotation)
     { 
     }
 }
