+++ GXPEngine/Stands/VegetableFruitStand.cs
@@ -7,7 +7,7 @@ namespace GXPEngine
 {
     class VegetableFruitStand : MarketStand
     {
-        public VegetableFruitStand(float givenX, float givenY, int givenRotation = 0) : base(givenX, givenY, "VegetableFruitsStand.png", givenRotation)
+        public VegetableFruitStand(float givenX, float givenY, float givenRotation = 0) : base(givenX, givenY, "VegetableFruitsStand.png", givenRotation)
         { 
         }
     }
